+++ Cargo.toml
@@ -13,7 +13,6 @@ members = [
     "crates/model_components",
     "crates/ns_lookup",
     "crates/util",
-    "crates/extensions",
     "tools/flightpublisher/",
     "tools/flightsubscriber/",
 ]
