+++ bin/spiced/Cargo.toml
@@ -15,7 +15,6 @@ spicepod = { path = "../../crates/spicepod" }
 app = { path = "../../crates/app" }
 runtime = { path = "../../crates/runtime" }
 flightrepl = { path = "../../crates/flightrepl" }
-extensions = { path = "../../crates/extensions" }
 tokio.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
