--- crates/extensions/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "extensions"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-exclude.workspace = true
-
-[dependencies]
-tracing.workspace = true
